+++ src/modules/classification/response/ai-post-list.dto.ts
@@ -1,0 +1,13 @@
+import { ApiProperty } from '@nestjs/swagger';
+import { Type } from 'class-transformer';
+import { AIPostServiceDto } from '../dto/getAIPostList.dto';
+
+export class AIPostListResponse {
+  @ApiProperty({ type: [AIPostServiceDto] })
+  @Type(() => AIPostServiceDto)
+  list: AIPostServiceDto[];
+
+  constructor(data: AIPostServiceDto[]) {
+    this.list = data;
+  }
+}
